core: block hot plug CPU with exclusive pinning
This patch prevent the user from hotplugging CPU when the VM is set with
exclusive pinning as we don't support it. Currently it will validated
only on the hot-plug command, which is less visible to the user.
With this patch, it won't be allowed and the user will be noticed.

Another addition is to consider within the validation the right pool of
shared CPUs to use.

Change-Id: Ie8d32b9e85879e6764bab63399c3d5762ba806db
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>