core: release CPUs
When we allocate CPUs we rely on the list we return back. This is
correct for the scheduling phase. But now we run the same logic on
canSchedule. There, when having a group of VMs, we use the cpuTopology
list, not the list we selected to allocate. The overtaken CPU remained
pinned and that harmed the canSchedule logic for VM groups.

When grouping the VMs, when the group is having an affinity set, it is
reordered. Therefore, it is required to sort them.

Change-Id: Idcae90fed953c4798f2c4bf8893bff0a1b7086ea
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>