Add deviceList parameter to connectStorageServer
Add a parameter to connectStorageServer, to send
information to the host with all device guids
(LUN wwids) that must be available on the host after
connecting, specifically for iSCSI storage.

This will prevent proceeding with a connection
before all the devices are available through
dm-multipath, which could lead to errors afterwards
(e.g., the VM not booting).

Bug-Url: https://bugzilla.redhat.com/2074029
Signed-off-by: Albert Esteve <aesteve@redhat.com>