webadmin: fix alignment on Resource Allocation table
Screenshot: https://imgur.com/a/CAIAnI1

This patch fixes the alignment under Resource Allocation > Disk
Allocation, when cloning a VM.
The disks' panel is being first created as part of the view initiation.
Determining the DC quota level (disabled/enforced) effects the number of
columns appear in the disk allocation view.
Since updating the quota happens later (not as part of initialization),
The columns' width can't be changed unless re-creating the disks lists
in the view, so the view needs to be created with the updated width.

Bug-Url: https://bugzilla.redhat.com/2081546
Signed-off-by: Shani Leviim <sleviim@redhat.com>