webadmin: Refactor NUMA for VM
With recent changes in the VM NUMA area the behavior
of the NUMA count and the NUMA support button was not consistent
and sometimes the changes were not saved on the backend at all.

This patch refactors the source code so that:
- The numaEnabled field has been removed from the UnitVmModel because
after the recent changes only the NUMA support button was affected by
its value. Instead, we enable and disable directly the NUMA support action.
- the NUMA support was renamed to NUMA pinning (in the code, not UI message) as
we now enable NUMA by setting the numaCount
- the updateNuma parameter sent to the backend is now calculated in the behaviors.
For the new VM, it is sent if the numa count > 0. For existing VMs, there is a series
of checks to determine if the NUMA should be updated on the backend. As a result,
the initialVmNumaNodes has been moved from UnitVmModel to the ExistingVmModelBehavior
because that is the only place where it is used.
- added support from configuring NUMA when creating a new VM from a template page

Bug-Url: https://bugzilla.redhat.com/2099225