core: change lock message of disks being downloaded
We changed the lock of disks that are being downloaded to be a shared
lock instead of an exclusive lock. Here we change the message that is
associated with this lock to be the same as the one we use for the
exclusive lock we take during upload-disk, which is more informative.