setup: Don’t fail cleanup on empty remote engine environment
When engine-setup fails on dnf repositories download, the followup
cleanup fails with

[ ERROR ] Failed to execute stage 'Clean up': 'NoneType' object has no attribute 'cleanup'

This patch moves the corresponding cleanup call from STAGE_CLEANUP to
STAGE_CLOSEUP.  STAGE_CLOSEUP only runs if we didn't fail and runs
non-critical stuff that we can't, or don't want to, rollback.  This
way, we can be sure that the cleanup is not run before the given
environment item is assigned.