core: permissive memory snapshot disk
When creating memory disk for live VM snapshot, the engine selects the
storage domain to use. The engine might select a storage domain that the
current user doesn't have permission on. This will fail to add the
memory disk and the snapshot operation will fail.
With this patch, the storage domains will be also filtered based on the
user permission.

Change-Id: Id1a1c529dde795628f214c0c9d9a28efac89eec0
Bug-Url: https://bugzilla.redhat.com/1565183
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>