engine: let backend decide when update numa
When updating a VM, it was required to send updateNuma
parameter that determined if the numa node changes will be saved.

The UI had to handle the logic which started to be quite complicated
- the numa nodes changes with memory, cpu, hugepages, numa count,
numa pinning, numa tune mode change.

This patch adds the possibility to leave out that parameter and
in that case the new numa node list is compared to the one
in the database. If they differ, the new numa node list will
be saved.