core: Uploading an image with an existing disk id
While uploading users can choose disk-id for thier uplaod and if
the disk-id was already existing the engine will make a new image
in lock state and fail the prosces. now it will check the disk-id before
creating and will fail if there is another disk-id with the same number.

Bug-Url: https://bugzilla.redhat.com/2095215
Signed-off-by: Artiom Divak <adivak@redhat.com>