core: set correct policy when reseting cpu pinning string
When we are importing a VM that is pinned to a host,
and the host is not present in the cluster where
the VM should be imported, the host pinning is cleared
as well as the cpu pinning string. However, the cpu pinning
policy was not updated and remained MANUAL causing the import
to fail when inserting the VM into the database.

This patch updates the cpu pinning policy to NONE when
cpu pinning string is cleared during import.

Bug-Url: https://bugzilla.redhat.com/2104115