core: update cpuTopology on status changes
Previously, once the CPU topology was initialized in
the VdsManager, it was not possible to change it without
the restart of the engine. There are 2 use cases:

1. When adding a new host, the manager is created, but no
CPU topology has been reported yet. As a result, the host
would act as if it does not support exclusive pinning until
the engine is restarted.
2. Changing the CPU topology of the host (not a common scenario,
but happens during development when using hosts in a virtual machine).

This patch updates the CPU topology in the VmManager if the state
of the host changes and the CPU topology has not been initialized yet.

Bug-Url: https://bugzilla.redhat.com/2104858