webadmin: Fix template import
The import of the template failed because one
of the provided listeners on cluster changes
was null. The listener was used to load OS into
a list for the VM, however, it is not needed
for the templates. This patch adds a check and
does not add the listener if it is null (as is
the case for the templates).

Bug-Url: https://bugzilla.redhat.com/2104597