ci: Checkout the PR head, not the github merge commit
In the CI github action, checkout the PR head commit instead of the
`actions/checkout@v2` default of the PR merge commit.  The practical
effect of this change is the exported artifacts snapshot suffix will
contain the commit id of the last commit on the PR instead of the
magic github generated merge commit.