engine: fix reboot on powering up VM
When the VM was powering UP and had the
next run configuration, the cold reboot did not
work. It had several reasons that have been fixed
in this patch:

1. the VM was just cleared based on getVmManager().getPowerOffTimeout()
2. the VmAnalyzer:proceedDisappearedVm did not have
RebootInProgress case and the default one was executed.
However, the code never reached the setColdRebootFlag() as
the VM was listed as asyncRunning and scheduled for rerun.
3. the next run config was never applied as the VM was present
in asyncRunning and never added to movedDown list.

Bug-Url: https://bugzilla.redhat.com/1912911