packaging: Make sure OVN certificates are renewed when needed
Due to incorrect staging and missing dependencies, OVN certificates
are not renewed if there are no other certificate renewals needed.
This patch fixes it by changing when OVN certificate info to PKI
entities is added.  It must be done:

- In STAGE_CUSTOMIZATION rather than STAGE_MISC.  STAGE_CUSTOMIZATION
  is where we ask for certificate renewal if needed.

- After the OVN plugin attributes are initialized so that the event
  condition doesn’t crash on a missing attribute.

- Before certificate renewal check is performed.

All this can be achieved by moving the single line of _misc_pki method
to _customization method.

It must be also done after
self.environment[oenginecons.PKIEnv.ENTITIES] is initialized but that
happens in STAGE_INIT, which is run early so no change is needed
here.  Note that we cannot add the OVN certificate info there in
STAGE_INIT because we don’t know yet whether OVN is used.

Bug-Url: https://bugzilla.redhat.com/2097558
Bug-Url: https://bugzilla.redhat.com/2097560