Turning procedures to functions
According to Postgres documentation a PROCEDURE executes SQL commands
and does not return any data.
In oVirt we are always returning a value (even VOID is considered as a
returned value), So we should use the term FUNCTION instead of PROCEDURE

This patch changes all relevant PROCEDURE declarations to a FUNCTION.
The patch is just a semantic change, was tested and can be safely
merged.

See also  https://bugzilla.redhat.com/2077794

Signed-off-by: Eli Mesika<emesika@redhat.com>