core: Prevent writing TPM device twice to OVA
TPM device has its own method writing its OVF entry.  But as a special
device, it’s written the second time to OVF among other devices.  When
such an OVF is imported, the VM has two TPM devices, which is not
allowed by libvirt and the VM fails to start.

This patch prevents writing the second TPM device entry to OVF.

Bug-Url: https://bugzilla.redhat.com/2112702