core: LiveMerge: move getVolumeInfo to executeCommand
Call to GetVolumeInfo was added to RemoveSnapshotSingleDiskLive
command before performing the steps, in order to check if the
Destination image (or the Top Image) exists on storage. It was added
to the procedure, which is determining the initial merge step during
the first run of performNextOperation.

GetVolumeInfo is a synchronous call, and calling it from
performNextOperation can block the callback thread which may block
other commands from making progress. This change moves the procedure
of setting the initial merge step back to executeCommand(), like it
is in ColdMergeSnapshotSingleDisk command