webadmin: warn when cloning snapshot with memory
When making a VM or a template from a snapshot with memory, the memory
is lost.  Moreover, the guest file system may be in an inconsistent
state without the memory.

This patch adds a confirmation dialog that is displayed
when user clicks Clone or Make Template buttons asking
if the user really wants to continue.

Bug-Url: https://bugzilla.redhat.com/1929376