core: skip disk profiles validation on memory disk
When we create the memory or metadata disks they are a result of other
command. These disks are not act as virtual disks. In some cases the
user can have permission to execute these command, but he may not have
the disk profile permission to add those disks.

This patch will skip setting the profile in those cases, allowing a user
to create attach disks for disks that are not a real virtual disks.

Change-Id: Ib04048d652c8d98df84ffccc4babd71730b218dc
Bug-Url: https://bugzilla.redhat.com/1565183
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>