core: Add support for zero-copy migrations
Zero-copy migrations help with migrating very large VMs (>1 TB) or
busy large VMs.

Zero-copy migrations are added using a new migration policy.  Since it
is a new feature, there may be unanticipated problems so the policy is
supposed to be used only when the already available migration policies
are insufficient to migrate a VM and the user is willing to accept the
risk.

Zero-copy can be used only with parallel migrations so it requires
parallel migrations enabled in Engine config.  If they are disabled in
VM/cluster configuration, they are still used if the zero-copy policy
is enabled.

Zero-copy migrations cannot be currently used with encrypted
migrations.  If encrypted migrations are requested, the given policy
can still be used but zero-copy is disabled.

The policy uses longer downtimes since it is aimed at very large VMs
and it is unlikely that short downtimes are any useful for such VMs.

It is supposed that Vdsm on the host supports zero-copy migrations,
since they are used only when parallel migrations are enabled,
i.e. with cluster version 4.7 by default, and it is supposed that the
hosts run an up-to-date OS.  If the host doesn’t support zero-copy
migrations then migration API violation will be logged there and
zero-copy won’t be enabled for the migration.

See also
https://ovirt.org/develop/release-management/features/virt/zerocopy-migrations.html.

Bug-Url: https://bugzilla.redhat.com/2089434