core: RemoveImage: handle failure in VDS command
RemoveImageCommand used to call for DeleteImageGroup VDS command, and
perform the required DB operations without polling on the Async Task.

One issue that it led to is when disk is configured with "Wipe on
Delete", and the wipe action took more than 50 hours (default max
async task polling time) - the task was interrupted, the image is not
deleted in VDSM, but removed from DB, not having any tracking for them
in the engine.

In this patch we add async task polling to RemoveImageCommand and
handle the removal of the image from the DB only if it succeeded,
otherwise we mark the image as ILLEGAL, and add an Audit Log, logging
the error and the IDs of the volumes that might need manual removal.
Additional removal of the now ILLEGAL disk will succeed, as the command
will not find any volumes associated with it in VDSM

Bug-Url: https://bugzilla.redhat.com/1836318