core: fail unomonitored backups
If the command monitoring a VM backup is no longer present in the
command_entities table, it will be impossible to finalize the backup and
it will be stuck in the finalizing phase.

This patch checks periodically for backups that are active but are no
longer monitored and fails them, notifying the user a cleanup may be
required.

Bug-Url: https://bugzilla.redhat.com/2107590
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>