core: create "MultipleStorageDomainsValidator" depending on the flow
Changing "createMultipleStorageDomainsValidator()" to be "protected"
in the "MoveOrCopyDiskCommand" class and overriding it in its derived
"LiveMigrateDiskCommand" class to add also the source Storage Domain ID.
Thus no need to check in the parent "MoveOrCopyDiskCommand" class
whether it's an LSM flow or not, in order to add also the source Storage
Domain ID to be checked by the created "MultipleStorageDomainsValidator".

Also removing the redundant "isDomainWithinThresholds()" destination
Storage Domain validation from the "LiveMigrateDiskCommand" class, since
it is correctly performed by the above "MultipleStorageDomainsValidator"
that validates both source & destination Storage Domains from its parent's
"MoveOrCopyDiskCommand::validateSpaceRequirements()".

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2116309