core: VDS Maintenance will wait for all disk transfers to happen and block new ones
if a user want to set a host to maintenance and at least one of
the disk is in transfer the command will fail and will send an error.
now it will not send an error it will wait for the disk transfer to finnish
(the host will be in preparingForMaintenance mode) and new transfers will be blocked from starting.

Bug-Url: https://bugzilla.redhat.com/2120228
Signed-off-by: Artiom Divak <adivak@redhat.com>