core: Refresh ISO cache when new guest agent installed
The latest available guest tools version is figured out by analyzing ISO
disks available on ISO and data storage domains. The list of available
ISO disks is cached and this cache is refreshed when user queries the
list of images on a storage domain. At that moment, the latest guest
tools version is refreshed and the guest agent installation hint
(presented to the user) is updated.

But ISO cache refresh does not happen after a new guest agent is
installed. That's why the guest agent installation hint is still shown
after installing the latest guest agent in the VM.

To fix that, this patch adds ISO cache refresh when the list of
applications for a particular VM is changed.

Change-Id: I6c25a97ae14cc40d073b3873c6bae9db3216204d
Bug-Url: https://bugzilla.redhat.com/1983401
Bug-Url: https://bugzilla.redhat.com/2120381
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>