core: vm will be down when active snapshot does not have memory
Previously, when trying-back/restoring a snapshot that doesn't have
memory while the client requested to restore memory, the VM would
have switched to Suspended state.

Now, the VM is set to Suspended only when the active snapshot, the
snapshot that was created by the trying-back/restore operation, is
set with memory volumes. Otherwise, the VM is set to Down.

See: https://github.com/oVirt/ovirt-web-ui/issues/1629

Signed-off-by: Arik Hadas <ahadas@redhat.com>