core: make register-vm a sync operation
In 38d5120a50c5f we've changed add VM from configuration to
be a blocking/sync operation. Here we do something similar for
register-vm, setting it to blocking/sync by default, for the
same reason - this operation doesn't involve copying the disks,
only add-lease is asynchronous and it is a relatively short
operation, so the caller won't wait much and at the end of the
call the VM is ready to be used.

This is particularly important for disaster recovery as we try
to start highly available VMs right after registering them. It
might have been better to change the disaster recovery scripts
to specify async=false, but that would require more efforts.

That change revealed an issue with the tests of ExportVmTemplate
since it is monitored by polling the job but tested as if it
was monitored by polling the VDSM tasks. Therefore the API
tests of ExportVmTemplate are also changed here.

Bug-Url: https://bugzilla.redhat.com/1968433
Signed-off-by: Arik Hadas <ahadas@redhat.com>