engine: Fix update cluster with numa VMs
When editing a cluster's CPU or version, an update
of the underlying VMs and templates is performed.
If any of them has numa enabled, the update cluster command
would fail because paramters's updateNuma = null is
changed to updateNuma = true in UpdateVmCommand.

Tnis patch makes it explicit that we are not updating
the numa by setting updateNuma = false parameter.