core: LSM: fix locking on restart
This patch fixes an issue when the engine is restarted during LSM,
the command gets stuck.

The cause of this issue was that EngineLocks were passed to
LiveMigrateDiskCommand, and from there to CreateSnapshotCommand in
the executeCommand() stage. There the locks are released, and
reacquired later on before snapshot remove phase. If the engine was
restarted during CreateSnapshot phase - both the CreateSnapshot and
LiveMigrateDisk commands resumed, running reacquireLocks, and then
before the Snapshot Remove phase, LiveMigrateDiskCommand attempted
to reacquire locks again, getting stuck.

In this patch we override reacquireLocks method in LiveMigrateDisk
command, doing nothing, and then the locks on Disk and VM are
acquired separately before the snapshot remove phase

Bug-Url: https://bugzilla.redhat.com/2110186