[NOT MERGE]ovn provider auth: use fips-compatible padding
Use OAEP with SHA256 and MGF1 padding in password encryption and
decryption during authenticating with ovirt-ovn-provieder on el9
deployments where FIPS is enabled.

This change should not be merged until all code paths using the modified
EngineEncryptionUtils facility are evaluated for correctness and until
this modification is made active only when FIPS is enabled.

Signed-off-by: Eitan Raviv <eraviv@redhat.com>
Change-Id: Ie895d2249716d6aafd64d79ae2d61083181b9703