core: trigger OvfDataUpdater when detaching a storage domain
This patch adds a call to OvfDataUpdater#triggerNow right after detaching
a storage domain in order to clean VMs/templates that were removed when
the storage domain was detached from the vm_ovf_generations table
immediately rather than in the next cycle we update the OVFs.

In 78d21eee4 we fixed an issue that caused importing VMs/templates with
old (stale) data. That fix could very well also fix the issue described
in the below mentioned bug, however, it is safer to also ensure that the
removed VMs/templates are also removed from the vm_ovf_generations table
to stop having old/stale data in the database that should never be used.

Bug-Url: https://bugzilla.redhat.com/1705338
Signed-off-by: Arik Hadas <ahadas@redhat.com>