engine: Re-generate CPU pinning
When migrating or suspending/resuming a VM with Resize
and Pin policy, the destination host's numa topology
might be different (e.g more CPUs in the numa node). While
the VM is able to run there, the CPU pinning might be different.

This patch re-generates the CPU pinning according to the numa
pinning for VMs with Resize and Pin policy when the VM has already
defined "current" CPU topology and "current" cpu pinning.

Bug-Url: https://bugzilla.redhat.com/2093948