Skip validation of build environment
We are checking maximum number of available file descriptors when
building engine for all support permutations. But in CBS this validation
doesn't work because maximum number of available file descriptors is not
reported properly. And we are also limiting number of workers for GWT
build, so we don't get near this limit.

Signed-off-by: Martin Perina <mperina@redhat.com>