frontend: Add TPM parameter when cloning a VM from a snapshot
When cloning a VM from a snapshot, isTpmEnabled parameter is not set
in VmSnapshotListModel, resulting in a loss of the TPM device and an
error if the TPM device is mandatory, i.e. for VMs with newer Windows
versions as the guest OS.  This patch fixes the omission and adds the
corresponding parameter.

Bug-Url: https://github.com/oVirt/ovirt-engine/issues/702
Bug-Url: https://bugzilla.redhat.com/2136180