core: Update TPM device in parameters
TPM device is not updated in
VmDeviceUtils.updateVmDevicesInParameters.  This results in omitting
the TPM device at least when exporting a VM to a data domain and an
error when the TPM device is mandatory, i.e. for VMs with a modern
Windows guest OS.  This patch fixes the omission and sets the
corresponding parameter.

Bug-Url: https://github.com/oVirt/ovirt-engine/issues/702