core: fix monitoring of RemoveVm
Previously, RemoveVm and RemoveAllVmImages didn't have a callback so they
didn't wait for all instances of RemoveImage to finish before finishing
RemoveVm and therefore the job of RemoveVm may have been completed while
the images are still locked in the database.

Now, RemoveVm is completed after RemoveImage completes and the images are
removed from the database.

Signed-off-by: Arik Hadas <ahadas@redhat.com>