core: Imported VM missing in OVFSTORE
When importing an OVA that is not originated in oVirt the VM will diappear
after detaching its storage domain. The vm will not be seen in the import VM
tab. From now on even if the imported VM is no originated in ovirt it could
be imported after detaching the storage domain.

Bug-Url: https://bugzilla.redhat.com/2028242
Signed-off-by: Artiom Divak <adivak@redhat.com>
Signed-off-by: Arik Hadas <ahadas@redhat.com>