core: prevent duplicate audit logs for ReduceImage
ReduceImage is an asynchronous operation. Previously, on positive flow we
logged a successful execution message on both 'execute' and 'end-action'
phases of the command. The former is incorrect and therefore dropped (we
typically log a message that says that the operation has started at this
point but we don't plan for another translation cycle so it's not added
as part of this change).

Bug-Url: https://bugzilla.redhat.com/2141437
Signed-off-by: Arik Hadas <ahadas@redhat.com>