core, webadmin: set new templates with specified vm-init
Previously, the backend (AddVmTemplate) set new templates with the vm-init
configuration of the VM that they were created from. This prevented users
from creating templates with modified vm-init configuration or without vm-
init configuration from VMs that are set with vm-init. Users could either
change the VM first or modify the created template to achieve this.

Now, the backend respects the vm-init configuration that is specified by
clients. While the handling of vm-init in REST-API works well, the UI
did not pass the vm-init configuration of the VM that templates were
created from. Therefore, the webadmin is changed to load the vm-init
configuration of the VM that a template is created from and pass it to
the backend now.

Signed-off-by: Arik Hadas <ahadas@redhat.com>