Revert "packaging: move selinux from ovirt-engine-common to ovirt-engine-setup"
This reverts commit e8eee2095731e98de7ee0051d3a0f66075d607e7.

Moving it causes grafana on a separate machine to fail, because the
plugin to enable httpd_can_network_connect is now only on the engine
machine. Reverting that patch should fix that.

This does not introduce a new regression, because the need for it was a
mere optimization, but is now not needed anymore, as we removed
use/integration of ansible-runner-service.

Change-Id: I35f57983c03c2de78b16ad63d37f399db103a07b
Bug-Url: https://bugzilla.redhat.com/2126778
Signed-off-by: Yedidyah Bar David <didi@redhat.com>