fix #779: propagate backup mode to template disks
Previously, all template disks were created with backup mode = NONE
regardless of the backup mode of the disks of the VM that the template
is created from.

Now the backup mode would propagate from the VM disks to the template
disks when their format is qcow2. If the template disks are created
as raw their backup mode would be NONE. Enabing to create template
disks with incremental backup enabled from VM disks that are raw is
outside of the scope of this change.

Signed-off-by: Arik Hadas <ahadas@redhat.com>