Lock devices during OVF export as this could cause a race otherwise
Closes: #797

During the creation of the ovf config export, oVirt does modifications
to the VmDevices.
Now another thread cleans those entries again, and if you are unlucky it
happens during the ovf generation itself.

DEBUG [org.ovirt.engine.core.vdsbroker.monitoring.VmDevicesMonitoring]
(EE-ManagedScheduledExecutorService-engineScheduledThreadPool-Thread-60)
[] VM 'xxxx' unmanaged device was marked for remove : {1}

And this causes the export to be skipped.

So we add a device lock during the ovf export to resolve this.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>